Misc fixes and changes
Fixed the bugged <<<<HEAD issue, shouldn't throw that error anymore.
I made the die's number public to the entire class. I am working on
getting the detective notes done.
I can try to start on the disprove logic, but I may not be able to test
it until the suggestion logic is done.