Floating/Sliding Windows for User Guide
So I used the white open space that was between the "End Turn" and
"Roll" button as the area this window will pop in an out.

Currently it has a button to Show it sliding in and a Hide button to
make it hide again.  This can be easily altered to fit our needs; we
need to decide at what points in the game it should pop open for the
user.